# api/openapi.yaml
openapi: 3.0.3
info:
  title: Time Ledger Sim
  version: 0.3.0
servers:
  - url: http://localhost:8080

paths:
  /healthz:
    get:
      summary: Health check
      responses:
        "200":
          description: OK

  /v1/version:
    get:
      summary: Build/version info
      responses:
        "200":
          description: Version info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionInfo"

  /v1/zones:
    get:
      summary: List zones
      responses:
        "200":
          description: Zones
          content:
            application/json:
              schema:
                type: object
                properties:
                  zones:
                    type: array
                    items:
                      $ref: "#/components/schemas/Zone"
                required: [zones]

  /v1/zones/{zone_id}/status:
    post:
      summary: Set zone status
      security:
        - AdminKey: []
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
        - name: X-Actor
          in: header
          required: false
          description: Operator name recorded as the audit actor (trusted only post-auth).
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetZoneStatusRequest"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "401":
          description: Unauthorized

  /v1/transfers:
    post:
      summary: Create transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferAppliedResponse"
        "202":
          description: Spooled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferSpooledResponse"
        "409":
          description: Idempotency conflict
        "503":
          description: Zone blocked

  /v1/balances:
    get:
      summary: List balances
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  balances:
                    type: array
                    items:
                      $ref: "#/components/schemas/BalanceRow"
                required: [balances]

  /v1/transactions:
    get:
      summary: List transactions
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/TransactionRow"
                required: [transactions]

  /v1/transactions/{transaction_id}:
    get:
      summary: Get transaction detail
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDetail"
        "404":
          description: Not found

  /v1/zones/{zone_id}/incidents:
    get:
      summary: List incidents for a zone
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Incidents
          content:
            application/json:
              schema:
                type: object
                properties:
                  incidents:
                    type: array
                    items:
                      $ref: "#/components/schemas/IncidentSummary"
                required: [incidents]

  /v1/incidents:
    get:
      summary: List recent incidents
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Incidents
          content:
            application/json:
              schema:
                type: object
                properties:
                  incidents:
                    type: array
                    items:
                      $ref: "#/components/schemas/IncidentSummary"
                required: [incidents]

  /v1/incidents/{incident_id}:
    get:
      summary: Get incident detail
      parameters:
        - name: incident_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentDetail"
        "404":
          description: Not found

  /v1/incidents/{incident_id}/action:
    post:
      summary: Apply incident action
      security:
        - AdminKey: []
      parameters:
        - name: incident_id
          in: path
          required: true
          schema: { type: string }
        - name: X-Actor
          in: header
          required: false
          description: Operator name recorded as the audit actor (trusted only post-auth).
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentActionRequest"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentDetail"
        "401":
          description: Unauthorized

  /v1/zones/{zone_id}/controls:
    get:
      summary: Get zone controls
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneControls"
    post:
      summary: Set zone controls
      security:
        - AdminKey: []
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
        - name: X-Actor
          in: header
          required: false
          description: Operator name recorded as the audit actor (trusted only post-auth).
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneControlsUpdate"
      responses:
        "200":
          description: Controls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneControls"
        "401":
          description: Unauthorized

  /v1/zones/{zone_id}/spool:
    get:
      summary: Get spool stats
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Spool stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpoolStats"

  /v1/zones/{zone_id}/spool/replay:
    post:
      summary: Replay spooled transfers
      security:
        - AdminKey: []
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
        - name: X-Actor
          in: header
          required: false
          description: Operator name recorded as the audit actor (trusted only post-auth).
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayRequest"
      responses:
        "200":
          description: Replay result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayResult"
        "401":
          description: Unauthorized

  /v1/zones/{zone_id}/audit:
    get:
      summary: List audit entries for a zone
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 120 }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEntry"
                required: [audit]

  /v1/sim/snapshot:
    post:
      summary: Export snapshot (admin)
      security:
        - AdminKey: []
      responses:
        "200":
          description: Snapshot
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized

  /v1/sim/restore:
    post:
      summary: Restore snapshot (admin)
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized

components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
    VersionInfo:
      type: object
      properties:
        service: { type: string }
        version: { type: string }
      additionalProperties: true
      required: [service, version]

    Zone:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string, enum: [OK, DEGRADED, DOWN] }
        updated_at: { type: string }
      required: [id, name, status]

    SetZoneStatusRequest:
      type: object
      properties:
        status: { type: string, enum: [OK, DEGRADED, DOWN] }
        reason: { type: string }
      required: [status]

    TransferRequest:
      type: object
      properties:
        request_id: { type: string }
        from_account: { type: string }
        to_account: { type: string }
        amount_units: { type: integer, format: int64, minimum: 1 }
        zone_id: { type: string }
        metadata: { type: object }
      required: [request_id, from_account, to_account, amount_units, zone_id]

    TransferAppliedResponse:
      type: object
      properties:
        status: { type: string, enum: [APPLIED] }
        transaction_id: { type: string }
        request_id: { type: string }
        created_at: { type: string }
      required: [status, transaction_id, request_id, created_at]

    TransferSpooledResponse:
      type: object
      properties:
        status: { type: string, enum: [SPOOLED] }
        spool_id: { type: string }
        request_id: { type: string }
      required: [status, spool_id, request_id]

    BalanceRow:
      type: object
      properties:
        account_id: { type: string }
        balance_units: { type: integer, format: int64 }
        updated_at: { type: string }
      required: [account_id, balance_units]

    TransactionRow:
      type: object
      properties:
        id: { type: string }
        request_id: { type: string }
        from_account: { type: string }
        to_account: { type: string }
        amount_units: { type: integer, format: int64 }
        zone_id: { type: string }
        created_at: { type: string }
      required: [id, from_account, to_account, amount_units, zone_id, created_at]

    PostingRow:
      type: object
      properties:
        account_id: { type: string }
        direction: { type: string, enum: [DEBIT, CREDIT] }
        amount_units: { type: integer, format: int64 }
      required: [account_id, direction, amount_units]

    TransactionDetail:
      type: object
      properties:
        id: { type: string }
        request_id: { type: string }
        from_account: { type: string }
        to_account: { type: string }
        amount_units: { type: integer, format: int64 }
        zone_id: { type: string }
        created_at: { type: string }
        metadata: { type: object }
        postings:
          type: array
          items: { $ref: "#/components/schemas/PostingRow" }
      required: [id, from_account, to_account, amount_units, zone_id, created_at, postings]

    IncidentSummary:
      type: object
      properties:
        id: { type: string }
        zone_id: { type: string }
        severity: { type: string, enum: [INFO, WARN, CRITICAL] }
        status: { type: string }
        title: { type: string }
        details: { type: object }
        detected_at: { type: string }
      required: [id, zone_id, severity, title, detected_at]

    IncidentDetail:
      type: object
      properties:
        id: { type: string }
        zone_id: { type: string }
        severity: { type: string, enum: [INFO, WARN, CRITICAL] }
        status: { type: string }
        title: { type: string }
        details: { type: object }
        detected_at: { type: string }
      required: [id, zone_id, severity, title, detected_at]

    IncidentActionRequest:
      type: object
      properties:
        action: { type: string, enum: [ACK, ASSIGN, RESOLVE] }
        assignee: { type: string }
        note: { type: string }
        reason: { type: string }
      required: [action]

    ZoneControls:
      type: object
      properties:
        zone_id: { type: string }
        writes_blocked: { type: boolean }
        cross_zone_throttle: { type: integer, minimum: 0, maximum: 100 }
        spool_enabled: { type: boolean }
        updated_at: { type: string }
      required: [zone_id, writes_blocked, cross_zone_throttle, spool_enabled]

    ZoneControlsUpdate:
      type: object
      properties:
        writes_blocked: { type: boolean }
        cross_zone_throttle: { type: integer, minimum: 0, maximum: 100 }
        spool_enabled: { type: boolean }
        reason: { type: string }

    SpoolStats:
      type: object
      properties:
        zone_id: { type: string }
        pending: { type: integer }
        applied: { type: integer }
        failed: { type: integer }
      required: [zone_id, pending, applied, failed]

    ReplayRequest:
      type: object
      properties:
        limit: { type: integer, default: 50 }
        reason: { type: string }

    ReplayResult:
      type: object
      properties:
        zone_id: { type: string }
        applied: { type: integer }
        failed: { type: integer }
      required: [applied, failed]

    AuditEntry:
      type: object
      properties:
        id: { type: string }
        actor: { type: string }
        action: { type: string }
        target_type: { type: string }
        target_id: { type: string }
        reason: { type: string, nullable: true }
        details: { type: object }
        created_at: { type: string }
      required: [id, actor, action, created_at]